pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chatboti"
version = "0.2.0"
description = "Raglet - RAG starter kit with FastAPI, MCP, and multiple LLM providers"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Thoughtworks Labs" }
]
keywords = ["rag", "retrieval-augmented-generation", "mcp", "fastapi", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.35.0",
    "ollama>=0.5.1",
    "path>=17.1.0",
    "pydash>=8.0.5",
    "fastapi>=0.116.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "aioboto3>=15.2.0",
    "mcp>=1.26.0",
    "numpy>=2.3.4",
    "slowapi>=0.1.9",
    "microeval>=0.3.1",
    "cyclopts>=3.0.0",
    "anyio>=4.12.1",
    "faiss-cpu>=1.9.0",
    "h5py>=3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "chromadb>=0.5.0",
    "pyarrow>=17.0.0",
    "faiss-cpu>=1.9.0",
]

[project.urls]
Homepage = "https://github.com/boscoh/chatboti"
Documentation = "https://github.com/boscoh/chatboti#readme"
Repository = "https://github.com/boscoh/chatboti.git"
Issues = "https://github.com/boscoh/chatboti/issues"

[project.scripts]
chatboti = "chatboti.cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["chatboti"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = "tests"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "asyncio: marks tests as async",
    "slow: marks tests as slow", 
    "integration: marks tests as integration tests"
]

